refactor(media): remove dead code and deduplicate batch/cover helpers
- album.py: drop return_exceptions=True and the dead result-check loop;
  _resolve_and_download already catches all exceptions internally so the
  gather never surfaces an Exception object
- media.py: add batch() as a static method on Media (the common base class)
- artist.py, playlist.py: remove their identical batch() copies; inherited
  from Media
- artwork.py: extract download_embed_cover() — wraps download_artwork() and
  discards the hi-res path, eliminating the duplicate _download_cover method
  that existed on both PendingSingle and PendingPlaylistTrack
- track.py, playlist.py: replace _download_cover with download_embed_cover;
  drop now-unused Covers imports
